selector time working
