+++ Php/conexion.php
@@ -2,7 +2,7 @@
 $host='localhost';
 $root='root';
 $password='';
-$nombrebd='medpriority_v2';
+$nombrebd='medpriority';
 $conn= new mysqli($host,$root,$password,$nombrebd);
 if(!$conn){
     die("No es posible conectar a la base de datos");
